packages/newsroom-core/pyproject.toml
  [project]
  name = "newsroom-core"
  version = "0.1.0"
  description = "Core library for verifiable newsroom"
  requires-python = ">=3.11"
  dependencies = ["pydantic>=2.7.0", "pyyaml>=6.0", "requests>=2.31.0", "typer>=0.12.0", "jsonschema>=4.22.0"]
  [project.optional-dependencies]
  dev = ["pytest"]